+++ Source/MiniSolarSystem/Public/Planet.h
@@ -142,7 +142,7 @@ public:
 	UFUNCTION(BlueprintCallable)
 	void BindSettingsIDs();
 
-	virtual int GetBodyRadius() const override;
+	virtual float GetBodyRadius() const override;
 
 #if WITH_EDITOR
 	virtual void PostEditChangeProperty(FPropertyChangedEvent& PropertyChangedEvent) override;
